Move commands into a separate file
