+++ etl-graph/utils.py
@@ -2,6 +2,7 @@ import json
 import subprocess
 from pathlib import Path
 from typing import List, Union
+
 from .config import *
